Revamp build logic for whole-project Javadoc
The previous approach used cross-project task access, which is strongly
discouraged and which actually broke when we converted the build from
Gradle+Groovy to Gradle+Kotlin.  The new approach implemented here uses
`Configuration`s to export relevant sets of Javadoc-related files and
paths from each subproject.  The root project's can build dependencies
on these so that the root `aggregatedJavadoc` task will have access to
the right material when needed.