Remove obsolete Travis CI material
We stopped using Travis CI nearly two years ago: #853 on January 2,
2021.  However, we still had a few leftover remnants of Travis-specific
logic.  Let's get rid of that obsolete material too.

Note that we previously had a few Rhino tests that we pretended to skip
only on Travis, but which we were actually skipping *everywhere* due to
setting `$TRAVIS` in the test environment.  Those tests are now
re-enabled, except for one which times out.  I'm not sure whether that
timeout is correct or a regression.  And in general, I'm not sure
whether we want to keep and re-enable these tests, keep them disabled,
or remove them entirely.