Preflight artifact publication during CI
We already have a convenient "fake" remote Maven repository that really
just sits in a local subdirectory under `build`.  Let's publish all
artifacts to this repository as part of the regular Ubuntu CI build.
That way we have an earlier opportunity to catch regressions that would
eventually interfere with the real artifact-publication process.