Avoid picking up `?` in file name
The
`com.ibm.wala.cast.js.rhino.callgraph.fieldbased.test.FieldBasedCGGamesTest.testProtoRPG`
test uses a URL with a `?` followed by a query string.  That previously
led us to try to create a filename containing `?`.  However, `?` is
disallowed in Windows file names, thereby causing this test to fail on
Windows.  The new logic uses only the part of the URL before the `?`,
which should be less likely to contain disallowed characters.