Simplify configuration of publication signatures
Previously we had two nearly identical publications: `local` and
`remote`.  `local` was unsigned, and was only intended for local
installation, but we could not really enforce that very well.
Conversely, `remote` was signed (for non-snapshots) and was only
intended for remote installation, which we likewise could not enforce
very well.

Now we just have a single `maven` publication.  We always attempt to
sign this publication's artifacts.  However, signature misconfiguration
is quietly allowed unless we are publishing a non-snapshot to a real
Maven remote repository.  Near as I can tell, this is the idiomatic way
to sign official remote releases but not sign local installations or
snapshot releases.

**Note:** [this wiki
documentation](https://github.com/wala/WALA/wiki/Creating-Releases-and-Snapshots#publishing-released-versions-locally)
should be updated after this commit is merged into `master`.  In fact,
that entire section can probably just be deleted:  after this change,
the obvious `publishToMavenLocal` target works as expected.  No
signatures are required when installing only into the local Maven
repository, even for non-snapshot releases.