Use Spotless for Java and Kotlin formatting
Spotless integrates nicely with the Java- and Kotlin-specific formatters
we were already using, and also with Gradle.  Furthermore, it *appears*
that Spotless may be ready for use with the Gradle configuration cache,
which isn't the case for the Google Java formatter plugin we were using
previously.  Thus, this change should move us a bit closer to #1206.