Add IntelliJ IDEA run configuration for testing configuration cache
This new "Build with configuration cache" run configuration performs
the same Gradle tasks as our GitHub Actions CI job when run on Ubuntu,
except that:

1. it excludes slow tests, and

1. it excludes tasks relating to native code and code signing that we
   already know are not compatible with the Gradle configuration
   cache.

This run configuration can be useful for identifying other
configuration-cache problems that we still need to work on.

Noe that this run configuration is not actually used by any GitHub
Actions CI job.  It's only here to be run manually, in an interactive
IntelliJ IDEA session.