Mark some kinds of tasks as configuration-cache-incompatible
If an incompatible task is scheduled to run, Gradle will warn about the
incompatibility, but will not fail the build.  Instead, the
configuration state will simply be discarded (not cached) when the build
completes.

This approach seems like a reasonable compromise given that we use some
standard Gradle tasks that are not yet compatible with the configuration
cache.  It lets us still use the cache for builds that don't involve
these incompatible tasks.