Be more strict when caching Gradle configuration
We are not yet ready to enable the Gradle configuration cache by
default.  If someone turns it on locally, though, we should use it in
[its stricter mode of
behavior](https://docs.gradle.org/current/userguide/configuration_cache.html#config_cache:stable).
Per the Gradle documentation, "It is recommended to enable
[`STABLE_CONFIGURATION_CACHE`] as soon as possible in order to be ready
for when we remove the flag and make the linked features the default."