Use testFixtures, not testArchives
`testArchives` was provided by a third-party plugin.  It made each
project's `test` classes and resources available as a `jar` file that
other targets could depend on.

`testFixtures`, on the other hand, is provided by the standard Gradle
`java-test-fixtures` plugin.  It creates a new `testFixtures` source set
distinct from `test` and `main`.  Each project's `testFixtures` classes
and resources are available as a `jar` file that other targets can
depend on.  Within a single subproject, `test` automatically depends on
`testFixtures`, and both `test` and `testFixtures` automatically depend
on `main`

The general idea is that `testFixtures` contains things that will be
used by `test` in one or more other subprojects, but which are not used
during standard execution and therefore do not belong in `main`.