Move `RandomGraph` into the `main` source set
Previously, `RandomGraph` was not being compiled at all, because it
was under `src/com/...`, rather than `src/main/java/com/...`.  We
overlooked this until now because nothing else actually uses
`RandomGraph`.  So I suppose we could delete this class entirely.  But
if we're going to keep it, then we should be compiling it so that we
know it has at least that level of validation.