Clean up Spotless format checks
I had not previously realized that Spotless offers a `kotlinGradle`
configuration separate from its `kotlin` configuration.  Most of WALA is
written in Java, with Kotlin only appearing in Gradle build scrips.  So
for most of WALA, we want to configure Spotless for `java` and
`kotlinGradle`, but not `kotlin`.  However, for the `build-logic`
subproject, we want both `kotlin` and `kotlinGradle`, since this
subproject uses Kotlin (and only Kotlin) for both its Gradle build
scripts and for the code that those scripts are building.

Explicitly configure the top-level `check`, `spotlessCheck`, and
`spotlessApply` tasks to depend on the same-named `build-logic` tasks.
In general, tasks implicitly depend on same-named tasks in subprojects.
But `build-logic` is an included build, not a subproject.  So for
`build-logic`, we need to set things like this up manually.

Remove Spotless checks of Kotlin code under `buildSrc`.  We no longer
use `buildSrc`, having already replaced it with `build-logic` as an
included build.