Prefix publications with `com.ibm.wala`
This used to happen implicitly, because `com.ibm.wala` was the root
project's name and `com.ibm.wala.` was a prefix for all subproject
names.  In #1247 we shortened the project name to just `wala` and
removed the common `com.ibm.wala.` prefix from all subproject names.  So
now we need to tweak things a bit to bring those prefixes back for
publications.

Tested locally using the `publishAllPublicationsToFakeRemoteRepository`
task, which actually just "publishes" to the local file system under
`build/maven-fake-remote-repository`.  The published files now consist
of:

* `com/ibm/wala/com.ibm.wala.*/maven-metadata.xml*`
* `com/ibm/wala/com.ibm.wala.*/1.6.1-SNAPSHOT/com.ibm.wala.*`
* `com/ibm/wala/com.ibm.wala.*/1.6.1-SNAPSHOT/maven-metadata.xml*`

That looks correct to me.

Fixes #1248.