Improve source locations for constructor calls for JS function statements (#1268)
This PR improves the source location used for the WALA IR instruction that performs a constructor invocation corresponding to a JS function statement. Previously, the location used was the location of the closest enclosing function / script, which was both too coarse and led to duplicate locations when there were multiple function statements in a scope. Now, we use the location of the function statement itself.

The implementation here is a bit hacky, but the source location code has complex invariants spread across multiple files and passes, so I think a local change is safest.