Migrate one test class from JUnit 4 to JUnit 5
The migration was mostly automated by IntelliJ IDEA.  We're starting
with just a single test class so that we have a clear demonstration of
the kinds of changes that will later be applied more broadly.

It's also good to verify here, by manual inspection, that we are
correctly running a mix of `core` tests from JUnit 5
(`ExtensionGraphTest`) and JUnit 4 (all the rest).  This confirms that
it's OK to use multiple `TestEngine`s at the same time, which in turn
confirms that it's OK to migrate tests from JUnit 4 to JUnit 5
incrementally rather than in one giant batch.