Remember to close a test jar file
The `instrumenter.nextClass()` call in this test helper method opens the
jar file that is identified by `instrumentedJarLocation`.  Previously,
nothing subsequently closed that jar file.  This open-file leak causes
trouble on Windows:  Windows refuses to delete an open file, which then
causes post-test `@TempDir` cleanup to fail.

Now we explicitly close `instrumenter`, which in turn closes the jar
file, which in turn allows post-test `@TempDir` cleanup to succeed on
Windows.