Tag additional slow tests
Empirically, these tests start but take a long time to finish on my
personal laptop.  Some of them wedge the entire laptop, requiring that I
cycle power to recover.  Granted, my laptop is a bit old and not the
most powerful.  I think it's reasonable to skip a few tests to make
development on this laptop feasible.  We still run the full test suite,
including "slow"-tagged tests, in our GitHub Actions workflows.