Suppress warnings about safe `this` leaks
Leaking `this` before the constructor has completed can be dangerous in
general, specifically if any methods are called or not-yet-initialized
fields are used.  However, in the specific cases addressed here, the
escaping `this` is merely recorded in fields of other objects for later
use.  Members of `this` instance are not actually used during
construction, so these escapes are safe.