Support building on JDK 21, and don't crash on JDK 21 bytecodes (#1339)
With these changes, WALA builds and runs on JDK 21, and our regression
tests pass, except for the Dalvik tests. The Dalvik test issue is
tracked as #1349. I think it's best to land these changes now, since
they include important fixes for JDK 21 support. We can then fix Dalvik
in a follow-up.

The main logic fix was to update the native models for
`java.lang.Thread` to account for implementation changes there. Also
there are a bunch of changes due to the deprecation of the `URL(String)`
constructor.