Update R8 and enable Dalvik tests on JDK 21 (#1353)
Updating to the latest R8 version lets us run the Dalvik tests on JDK
21. We disable one check due to
https://issuetracker.google.com/issues/316744331; we'll have to update
R8 again once that is fixed. We also needed to tweak how we construct
the test jar for use with Dalvik. My best guess is some classes caused
issues with D8 before, but they no longer do so. Finally, we performed
some cleanup in `JVMLDalvikComparisonTest` to make failures easier to
understand.

Fixes #1349