Upgrade to Eclipse 4.30.0
This newer Eclipse release requires Java 17 or better.  However, we
still prefer to keep most of WALA compatible with Java 11.  Therefore,
this Eclipse upgrade adds nontrivial build infrastructure to let us use
a newer Java release only where Eclipse dependencies require it.

A new `com.ibm.wala.jdk-version` property in `gradle.properties` allows
setting the default Java version used throughout WALA.  If this property
is set to 17 or higher, then that same version is used both for
Eclipse-dependent and Eclipse-independent WALA components.  However, if
this property is set to 16 or earlier, then it is only used for WALA's
Eclipse-independent components; the Eclipse-dependent components will
override this and use Java 17 instead.

In a default configuration, most of WALA will be compiled with (and
for) Java 11, with the Eclipse-dependent components using Java 17
instead.

Resolves #1354.