Use an explicit setter to help IntelliJ IDEA
For some reason, IntelliJ IDEA's Kotlin support does not recognize
that `destinationDir = ...` is a correct invocation of the
`setDestinationDir` setter.  That works just fine when run in Gradle,
but IntelliJ IDEA flags it as an error.  To keep IntelliJ IDEA happy,
we can write this property assignment as an explicit setter call
instead.