Switch to ConcurrentHashMap in SeqClassHierarchyFactory (#1379)
I believe this should address #1377. See
https://github.com/wala/WALA/commit/74b6c744926146ecaeaae5da5a74e18c95891fff.
For some reason at that point we decided not to use the type system to
enforce use of a `ConcurrentHashMap` always. In this spirit of trying
not to break things I didn't do that here, but instead changed
`SeqClassHierarchyFactory` to use `ConcurrentHashMap`.

Fixes #1377 (though we can reopen if the crash still happens)