Remove snapshot uploading CI job (#1410)
This broke recently since I had to rotate my Sonatype credentials. I
think we should just disable it. Building a snapshot locally is pretty
easy, and it's generally not a great practice to add a dependence on a
remote snapshot build that may change unexpectedly. If anyone needs us
to cut a release to get some change, they can just ask and we will do
it.