Prefer Gradle Sync tasks over Copy tasks
`Sync` tasks remove any files in the destination that were not actually
copied from the source.  This feels safer than `Copy`, which is happy
to retain leftover destination files.  Those leftover files might have
been copied manually or by an earlier version of the build logic, either
of which threatens reproducibility of a build.