Support building and running on JDK 22 (#1419)
We don't immediately crash on JDK 22 bytecodes, but new bytecode
features have not been tested.

We modify the `TestList` test input to avoid a blowup when running
`SlicerTest.testList`.
https://github.com/openjdk/jdk/commit/b62e774e6a531db934de04211724a2a8159d94db
introduced calls to `String.format` from `Integer.parseInt`, which
dramatically increased the number of reachable methods for the previous
version of this test, which in turn led to very slow performance with
the slicer. Changing the test to use a `List<String>` and string
constants removes this blowup. (Whole-program analysis is fun!)

Fixes #1414