Update and broaden use of Java toolchain resolver
Use the latest 0.8.0 release of this plugin.

Use this plugin in the build-logic subproject, in case that subproject
needs a Java toolchain that is not already installed system-wide or by
the user.  In particular, this subproject compiles Kotlin for Java 11,
but there's no guarantee that a Java 11 toolchain is already present.