refactor: remove getTestName from IRTests.java so tests do not rely on method name (#1485)
Cleaned up `ECJJava15IRTest.java` by refactoring the tests to be
parameterized to not rely on method name for running the tests. Also
removed the `getTestName` from `IRTests` so the tests do not rely on
that method to get the name of the test it should run. With this we pass
in the name and tests can now be parameterized as well instead of
relying on naming tests differently

---------

Co-authored-by: Manu Sridharan <msridhar@gmail.com>