Simplify JaCoCo report aggregation
When aggregating, automatically include every subproject that uses the
`jacoco` plugin.  The logic for doing this works out quite nicely thanks
to `PluginContainer::withId`.  That method executes an action
immediately if a project has already loaded the plugin of interest, _or_
registers the action to run later if the project loads the plugin of
interest later.  Thus, everything works as intended regardless of the
order in which subprojects are configured.