Upgrade to Gradle 9.0.0-RC1
This change uses a Gradle release candidate, so I don't intend to
merge it into the `master` branch.  However, it demonstrates that we
are as ready for Gradle 9 as we should be for now.

Note that Gradle 9 requires Java 17 or newer to run.  It can still use
older JVMs for compiling or running other tasks, though.  Thus, we can
still build and distribute JARs that other people can use in Java 11
JVMs.

Fixes #1522.