Hide Eclipse version management from root project
Previously the root Gradle project determined the version of Eclipse
JARs that all Eclipse-integrated WALA components should use.  Now we
hide that deeper inside WALA internals, as part of the existing
`WalaMavenCentralReleaseConfigurerExtension` class.  The intent is to
make that class more self-contained, and to reduce coupling between
subprojects and the root project.