Simplify rpath handling for Gradle native linking
Previously, we used various helper functions to add an rpath for each
shared library.  However, it turns out to be simpler to manage rpaths
all in one place.  Now, we wait until linker arguments are actually
needed, and then compute _all_ required rpath arguments based on the
_complete_ set of linked libraries.

A nice side effect of this change is that preparing to add these rpaths
can work solely on a linker task _provider_, without requiring that the
linker task be eagerly instantiated.