Treat `addJvmLibrary` as a `CppBinary` method
Previously, this helper was an `AbstractLinkTask` extension method.
However, we no longer do anything special with that link task.  Instead,
we're operating on the enclosing `Project` instance and the target
`CppBinary` instance.  So we can simplify a bit by turning
`addJvmLibrary` into a `CppBinary` extension method instead of an
`AbstractLinkTask` extension method.

A desirable side effect of this change is that we no longer force eager
configuration of an `AbstractLinkTask` task in order to add the JVM
library to its target binary.