Avoid cascaded project property lookups
Gradle subprojects inherit properties from their parent projects.  Thus,
looking for a property on a project may involve a cascaded sequence of
lookups, eventually reaching the root project.  Unfortunately, these
cascades are incompatible with Gradle project isolation.  So we now use
a different strategy for property lookups.  This strategy should give us
the same results in the case of properties on the root Gradle project,
which is all we really care about.  However, this strategy is
(apparently) compatible with project isolation.

Before this change, trying to enable Gradle project isolation failed
with 632 problems, in 10 distinct locations, having 82 distinct
messages.  After this change, trying to enable Gradle project isolation
fails with 451 problems, in 7 distinct locations, having 58 distinct
messages.  The problems that remain appear to all be due to third-party
Gradle plugins that are not yet compatible with project isolation.  See
<https://github.com/orgs/wala/projects/4> for links to existing bug
reports for some (but not all) of those incompatibilities.