Fix some weird indentation
When an `if` or `else` has a `//` comment on the same line, and the
conditionally executed code is just a single statement, then Spotless
uses rather strange indentation:

```java
if (condition) // comment
statement;
```

Let's wrap each such conditionally executed statement in a block so that
the indentation is more reasonable:

```java
if (condition) {
  // comment
  statement;
}
```

I haven't applied this change to _all_ conditionally executed non-block
statements.  I only changed those with a `//` comment that led to the
weird indentation shown above.