Improve modeling and caching of a Gradle task
The `:cast:java:ecj:run` task passes a directory full of source files as
a command-line parameter to a subprocess.  We should probably assume
that everything in that directory is a potential input to the task.

This same task can be cached, now that we have its inputs and outputs
modeled correctly.  It takes about three seconds to run, but its
outputs are small.  So we'd rather grab cached outputs if possible
instead of computing them from scratch.