Prefer `Path` for filesystem paths
A `String` can represent just about anything.  For strings that
represent paths in the filesystem, prefer to use `Path`.  It gives us a
better API for file-related operations, makes argument-swapping errors
less likely, and is simply a semantically richer description of what we
are doing.

Ordinarily I would consider an API change like this as
backward-incompatible and therefore requiring careful management of
version numbers, deprecation, etc..  But the APIs affected by this
change are all in test code.  So I think we can justify making this
change without a lot of backward-compatibility ceremony.