Revamp utility for managing sets of class names
Previously the main implementation of this utility was `FileOfClasses`.
This implementation was built around an assumption that it would read a
sequence of patterns from an `InputStream`.  Unfortunately, there are a
few places where the code preferred to provide patterns in some other
way.  That code was forced to simulate an `InputStream` using tools like
`ByteArrayInputStream`.  So we might have a sequence of patterns, but we
had to concatenate them into a single newline-separated `String` just so
that `FileOfClasses` could parse that monolith and break it up into
newline-separated patterns again.  Unnecessarily cumbersome!

The replacement for this class is `PatternsFilter`.  One can build a
`PatternsFilter` instance from an `InputStream` as before.  But one can
also build one from a `Path`, a `File`, or a `Stream<String>`.  The
latter offers an easier API for callers that already have their patterns
separated out, e.g., in an array or `List`.

`PatternsFilter` also offers a builder interface that can be useful if
patterns are to be added in several operations rather than a single
constructor call.

Previously, `FileOfClasses` allowed new patterns to be added at any
time.  It went to some extra effort to lazily (re)build the `Pattern`
that represented all collected patterns when needed.  However, in
practice, nothing in WALA ever added new patterns after some initial
construction phase.

The new `PatternsFilter` class is immutable:  patterns are passed to the
constructor and cannot change thereafter.  The builder allows patterns
to be added in several steps, of course:  that's what it is for!  But
once the builder is used to actually build an instance, no further
changes are allowed.  This immutability restriction greatly simplifies
management of the aggregated regular expression, which in turn makes the
membership check method simpler and faster.  Immutability also makes
`PatternsFilter` trivially thread-safe after construction, which was not
true for `FileOfClasses`.

The old `FileOfClasses` class is kept around for now, but is marked as
deprecated.  The same goes for its `SetOfClasses` base class.  The new
version of the latter is `StringFilter`:  a more appropriate name given
that there's really nothing specific to _classes_ about any of this
code.