JetBrains annotations are not needed at run time
The `@Language` annotation that we now use has `CLASS` retention
policy, not `RUNTIME`. That means that the annotation is retained in
the compiled class file but is not available for reflection at run
time by the JVM.  Thus, there's no need to have this annotation's
implementing JAR available as a run-time dependency.