Remove two useless calls to getter methods
The call to `NewInstruction.getType()` does not do anything useful.
`NewInstruction` is a `final` class, so we know exactly which
implementation of `getType()` will be called.  That implementation has
no side effects.  So if we don't need the returned value, then there's
no reason to call it.

The call to `super.isAssigned()` _appears_ to be useless as well.  We
know exactly which implementation of `isAssigned()` will be called.
That implementation has no side effects.  So if we don't need the
returned value, then there's no reason to call it.  **However,** I do
wonder whether this was supposed to be a call to `super.setAssigned()`
instead.  Manu, any thoughts on that?