Remove an unused package-scoped method parameter
The package-scoped overload of `SpecializedInstantiator.createInstance`
had a `Set<? extends SSAValue> seen` formal parameter.  That parameter
would be set to a new `HashSet` instance if initially `null`, but there
were no other uses of this parameter in the method.  We may as well
discard that parameter entirely.  Backward-compatibility should not be
a problem, since this method was previously visible only to code in the
`com.ibm.wala.dalvik.ipa.callgraph.androidModel.parameters` package.