Suppress warnings for seemingly unused catalog entries (#1650)
Each of these entries is actually used. But those uses are in
`build-logic` code that must look these entries up as strings instead of
using generated Kotlin bindings. That causes IntelliJ IDEA to mistakenly
think that these entries are unused.

We could turn off that IntelliJ IDEA warning entirely, for the whole
catalog. But the vast majority of catalog entries are used in standard,
detectable ways. It's still useful to have this warning in case any of
those stop being used in the future.