Factor out setting `$JAVA_HOME` for subprocesses
If a Gradle `Exec` task runs a subprocess that uses a JVM, that
subprocess may need to know where to find a working Java installation.
We know where to find the Java installation that is already running
Gradle, so that's generally suitable for subprocesses to use too.

The logic to set `$JAVA_HOME` for an `Exec` task is not _terribly_
complicated, but it's not trivial either.  Let's extract it into an
`Exec` extension method for easy reuse.  Right now there's just one such
use, but I plan to add another soon in a separate commit.