Treat `android.jar` as test data, not code
Previously we included the directory containing `android.jar` in the
classpath for `dalvik` tests, so that
`com.ibm.wala.dalvik.test.util.Util.androidJavaLib()` can find it using
standard resource loaders.  However, this one file is the _only_
component of an Android SDK installation that we use during testing.
Therefore, it is more proper to treat just the `android.jar` file as a
test resource, without picking up everything else in that same
directory.